fix categorization of level2 and reading csv data
One problem was that missing values in level 2 would result in 0 levels for these variables. a level 2 variable has to have 1 level for each group. I now treat 0 levels as 1.

The other problem was that the encoding of csv files was not taken into account. Windows user often have iso-8859, so I try to detect the encoding first.