update tests due to changes for output table
Note that previously the intercept of the beta was not 0, now it is. I assume there were some problems in calculating betas; maybe they are still present, but this at least looks better than previously. So I will just update the tests, without inquiring why the values were so strange in the first place. The unstandardized estimates have not changed, this is more important anyway.