add non-reactive mode
For long computations it makes sense to isolate the inputs and use an action button to start the calculation. This is now possible by adding a non-reactive mode where all inputs are isolated. By default this mode is not on. To turn it on a checkbox needs to be unchecked, which makes a button visible for calculation and isolates all inputs.